Combine script unit tests
Move all the unit tests from `ownded` and `borrowed` to the newly
created `tests` module.

Done to assist easily reading all the script units tests.
