+++ primitives/src/script/borrowed.rs
@@ -205,78 +205,3 @@ delegate_index!(
     RangeToInclusive<usize>,
     (Bound<usize>, Bound<usize>)
 );
-
-#[cfg(test)]
-mod tests {
-    // All tests should compile and pass no matter which script type you put here.
-    type Script = super::super::ScriptSig;
-
-    use alloc::{borrow::ToOwned, vec};
-
-    #[test]
-    fn script_from_bytes() {
-        let script = Script::from_bytes(&[1, 2, 3]);
-        assert_eq!(script.as_bytes(), [1, 2, 3]);
-    }
-
-    #[test]
-    fn script_from_bytes_mut() {
-        let bytes = &mut [1, 2, 3];
-        let script = Script::from_bytes_mut(bytes);
-        script.as_mut_bytes()[0] = 4;
-        assert_eq!(script.as_mut_bytes(), [4, 2, 3]);
-    }
-
-    #[test]
-    fn script_to_vec() {
-        let script = Script::from_bytes(&[1, 2, 3]);
-        assert_eq!(script.to_vec(), vec![1, 2, 3]);
-    }
-
-    #[test]
-    fn script_len() {
-        let script = Script::from_bytes(&[1, 2, 3]);
-        assert_eq!(script.len(), 3);
-    }
-
-    #[test]
-    fn script_is_empty() {
-        let script: &Script = Default::default();
-        assert!(script.is_empty());
-
-        let script = Script::from_bytes(&[1, 2, 3]);
-        assert!(!script.is_empty());
-    }
-
-    #[test]
-    fn script_to_owned() {
-        let script = Script::from_bytes(&[1, 2, 3]);
-        let script_buf = script.to_owned();
-        assert_eq!(script_buf.as_bytes(), [1, 2, 3]);
-    }
-
-    #[test]
-    fn test_index() {
-        let script = Script::from_bytes(&[1, 2, 3, 4, 5]);
-
-        assert_eq!(script[1..3].as_bytes(), &[2, 3]);
-        assert_eq!(script[2..].as_bytes(), &[3, 4, 5]);
-        assert_eq!(script[..3].as_bytes(), &[1, 2, 3]);
-        assert_eq!(script[..].as_bytes(), &[1, 2, 3, 4, 5]);
-        assert_eq!(script[1..=3].as_bytes(), &[2, 3, 4]);
-        assert_eq!(script[..=2].as_bytes(), &[1, 2, 3]);
-    }
-
-    #[test]
-    fn encode() {
-        // Consensus encoding includes the length of the encoded data
-        // (compact size encoded length prefix).
-        let consensus_encoded: [u8; 6] = [0x05, 1, 2, 3, 4, 5];
-
-        // `from_bytes` does not expect the prefix.
-        let script = Script::from_bytes(&consensus_encoded[1..]);
-
-        let got = encoding::encode_to_vec(script);
-        assert_eq!(got, consensus_encoded);
-    }
-}
