+++ primitives/src/script/owned.rs
@@ -211,140 +211,3 @@ impl<'a, T> Arbitrary<'a> for ScriptBuf<T> {
         Ok(Self::from_bytes(v))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    // All tests should compile and pass no matter which script type you put here.
-    type ScriptBuf = super::super::ScriptSigBuf;
-
-    use alloc::string::ToString;
-    use alloc::vec;
-    #[cfg(feature = "std")]
-    use std::error::Error as _;
-
-    use super::*;
-
-    #[test]
-    fn script_buf_from_bytes() {
-        let bytes = vec![1, 2, 3];
-        let script = ScriptBuf::from_bytes(bytes.clone());
-        assert_eq!(script.as_bytes(), bytes);
-    }
-
-    #[test]
-    fn script_buf_as_script() {
-        let bytes = vec![1, 2, 3];
-        let script = ScriptBuf::from_bytes(bytes.clone());
-        let script_ref = script.as_script();
-        assert_eq!(script_ref.as_bytes(), bytes);
-    }
-
-    #[test]
-    fn script_buf_as_mut_script() {
-        let mut script = ScriptBuf::from_bytes(vec![1, 2, 3]);
-        let script_mut_ref = script.as_mut_script();
-        script_mut_ref.as_mut_bytes()[0] = 4;
-        assert_eq!(script.as_mut_bytes(), &[4, 2, 3]);
-    }
-
-    #[test]
-    fn script_buf_into_bytes() {
-        let bytes = vec![1, 2, 3];
-        let script = ScriptBuf::from_bytes(bytes.clone());
-        let result = script.into_bytes();
-        assert_eq!(result, bytes);
-    }
-
-    #[test]
-    fn script_buf_into_boxed_script() {
-        let bytes = vec![1, 2, 3];
-        let script = ScriptBuf::from_bytes(bytes.clone());
-        let boxed_script = script.into_boxed_script();
-        assert_eq!(boxed_script.as_bytes(), bytes);
-    }
-
-    #[test]
-    fn script_buf_capacity() {
-        let script = ScriptBuf::with_capacity(10);
-        assert!(script.capacity() >= 10);
-    }
-
-    #[test]
-    fn script_buf_reserve() {
-        let mut script = ScriptBuf::new();
-        script.reserve(10);
-        assert!(script.capacity() >= 10);
-    }
-
-    #[test]
-    fn script_buf_reserve_exact() {
-        let mut script = ScriptBuf::new();
-        script.reserve_exact(10);
-        assert!(script.capacity() >= 10);
-    }
-
-    #[test]
-    fn script_buf_default() {
-        let script: ScriptBuf = ScriptBuf::default();
-        assert!(script.is_empty());
-    }
-
-    #[test]
-    fn script_consensus_decode_empty() {
-        let bytes = vec![0_u8];
-        let mut push = bytes.as_slice();
-        let mut decoder = ScriptBuf::decoder();
-        decoder.push_bytes(&mut push).unwrap();
-
-        let got = decoder.end().unwrap();
-        let want = ScriptBuf::new();
-
-        assert_eq!(got, want);
-    }
-
-    #[test]
-    fn script_consensus_decode_empty_with_more_data() {
-        // An empty script sig with a bunch of unrelated data at the end.
-        let bytes = vec![0x00_u8, 0xff, 0xff, 0xff, 0xff];
-        let mut push = bytes.as_slice();
-        let mut decoder = ScriptBuf::decoder();
-        decoder.push_bytes(&mut push).unwrap();
-
-        let got = decoder.end().unwrap();
-        let want = ScriptBuf::new();
-
-        assert_eq!(got, want);
-    }
-
-    #[test]
-    fn decoder_full_read_limit() {
-        let mut decoder = ScriptBuf::decoder();
-        // ByteVecDecoder length prefix is CompactSize: needs 1 byte.
-        assert_eq!(decoder.read_limit(), 1);
-
-        // Script length prefix = 32.
-        let mut push = [32_u8].as_slice();
-        decoder.push_bytes(&mut push).unwrap();
-        // Limit is 32 for the script data.
-        assert_eq!(decoder.read_limit(), 32);
-
-        // Provide 1 byte of script data decreasing the read limit by 1.
-        let mut push = [0xAA_u8].as_slice();
-        decoder.push_bytes(&mut push).unwrap();
-        assert_eq!(decoder.read_limit(), 31);
-    }
-
-    #[test]
-    #[cfg(feature = "alloc")]
-    fn decoder_error_display() {
-        let bytes = vec![0x01_u8];
-        let mut push = bytes.as_slice();
-        let mut decoder = <ScriptBuf as Decodable>::Decoder::default();
-        decoder.push_bytes(&mut push).unwrap();
-
-        let err = decoder.end().unwrap_err();
-        assert!(!err.to_string().is_empty());
-        #[cfg(feature = "std")]
-        assert!(err.source().is_some());
-    }
-}
